load: import all platforms
Currently if you try to import a non-native image the blobs will get
GC'd and unpack fails with with "not found" on the blobs.

Import doesn't really have a way to select which platforms to important,
so I don't think it makes sense to make this optional. After all the tar
could just be modified with the desired platforms.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>