rootless: allow loading an existing AppArmor profile
nerdctl run
---
Rootless nerdctl now applies the `nerdctl-default` profile to containers by default,
if it is already loaded with `sudo nerdctl apparmor load`.

Still defaults to "unconfined" when the profile is not loaded.

nerdctl info
---
Now `nerdctl info` shows "apparmor" when the AppArmor kernel module is enabled,
regardless to whether the `nerdctl-default` profile is loaded or not.
When the profile is not available, `nerdctl info` shows a warning message.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>