ocihook: Add a networking namespace annotation key
With VM based runtimes (e.g. Kata), the spec.State.Pid value does not
necessarily runs in the container/pod networking namespace, but can also
be in the host one.

With those runtime, using the passed Pid to resolve the networking
namespace for OCI plugins to be used result in setting the container
network entirely in the host namespace. We add a
"nerdct/network-namespace" for runtimes to explictly tell nerdctl which
netns path to use instead of deriving it from the runtime PID.

Signed-off-by: Samuel Ortiz <s.ortiz@apple.com>