Adding --attach flag to start
This change adds the -a/--attach flag to attach to stdout/stderr
and os events. It follow the style used in run without -d to do
this.

One decision was if the user uses the -a flag and there is a log
label to choose -a and warn that the label is ignored.

Signed-off-by: Matt Farina <matt.farina@suse.com>