Prevent double-creation of nerdctl default network.
This patch adds logic to pkg/netutil to prevent nerdctl from
accidentally creating a second default network if a non-nerdctl-created
network with the same default name already happens to exist.

Additionally, a new `nerdctl/default-network` label definition has been
added to indicate whether a network was created by nerdctl with the
express purpose of being used as a default network.

Signed-off-by: Nashwan Azhari <nazhari@cloudbasesolutions.com>