setPlatformOptions: Change signature
This changes `setPlatformOptions`' signature to avoid passing back a fresh
copy of an internalLabels structure when we could just pass in a pointer
and have the function set whats needed. This additionally removes the
opts parameter as the naming is a bit odd for this. To me, reading the
function name it'd seem as if we're going to make use of the opts in the
body (e.g. actually apply/set them) but we just append to the slice
internally. I've changed to returning a new slice that makes it more obvious
that all we're doing internally is tallying up each platforms specific
options and returning them.

Signed-off-by: Danny Canter <danny@dcantah.dev>