Add dbus-user-session related files
In Photon OS or in Redhat derivatives, there is no dbus-user-session
package. It is better to keep them in the repo and users can use these
files on the need basis on cgroup v2 systems.

I have tested this on Photon OS with 6.x kernel and cgroup v2 enabled.

Update the FAQ for the same.

Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>