create service containers not detached by default
service parsing uses a utility function `newContainer` to create
container objects. `compose up` and `compose run` are meant to create
containers as not detached by default, but this function was explicitly
setting the detached option on new containers to true, causing stderr IO
from issues in `compose up` to be suppressed.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>