Fix compatibility issue with Harbor token service
When a request contains `Authorization` header but fails to
authenticate, Harbor registry returns `Basic realm="harbor"` as the
authentication challenge [1], instead of the token service configured.

nerdctl caches pull-scoped tokens for later use, so when a *push*
request is sent with the cached pull-scoped token, Harbor denies the
request and instructs the client to attempt "Basic" authentication, which
consistently fails if "Basic" authentication is disabled or reserved for
special accounts only.

`nerdctl push` to a Harbor registry could fall into this "Basic"
authentication loop and eventually fail with 401 Unauthorized.

This patch detects such pattern of repeated requests and retries the
request without `Authorization` header to get a proper "Bearer"
challenge from the registry.

[1]: https://github.com/goharbor/harbor/blob/6113469a567623054a58d0f5506b90a55adca7ff/src/server/middleware/v2auth/auth.go#L95-L98

Fixes: https://github.com/containerd/nerdctl/issues/1675
Signed-off-by: Lin Dongyu <l2dy@icloud.com>