Fix incorrect usage of fmt.Fprintf()
In quite a few places, we take the pattern of

  b := <template.Execute() output>
  fmt.Printf(..., b.String() + "\n")

This means we end up passing arbitrary strings as a printf-style
formatting string, leading to issues whenever there is a literal `%` in
the output.  As in #2209, convert to fmt.Fprintln() as appropriate so
that we emit the rendered template as-is.

Signed-off-by: Mark Yen <mark.yen@suse.com>