Add SOCI index digest flag
`soci image rpull` and the SOCI library both allow for a particular
SOCI index to be specified, so this change adds support for passing this
in. Since no existing structure was in place to pass flags to remote
snapshotters, this change also introduces the structures needed to
support such future changes.

Signed-off-by: David Son <davbson@amazon.com>