Remove "soci image rpull" from testing suite
Since SOCI plans on removing this functionality from the CLI, we can
preemptively remove our test's dependency on this component of the CLI
by hardcoding the expected number of mounts per image. This is fine
because the same image with the same SOCI config will always have the
same amount of mounts. The only caveat is that one must know how many
mounts to expect from a certain image.

Additionally, small refactoring was done to allow for more future test
coverage, and to allow TestSociRun to be as flexible as TestSociPull.

Signed-off-by: David Son <davbson@amazon.com>