Rationalise build constraints for Go 1.19 onwards
This mostly involves replacing anything that was trying to be "Not
Windows" with the new-in-Go-1.19 "unix" build-constraint. Effectively,
anything that contained both "linux" and "freebsd" but not "windows", or
excluded only "windows", is now "unix".

A couple of needless constraints were removed when the filenames already
carried the appropriate constraint.

A handful of files were renamed, so that now "unix"-suffixed files all
use the "unix" constraint, and "other" is used when the constraint is
more-complex, for "no specific implementation" cases.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>