Add SOCI handler wrapper
As part of upgrading nerdctl to containerd v2, we must update
dependencies to depend on v2 as well. As SOCI maintainers do not wish to
upgrade to containerd v2 yet, this is a workaround to avoid depending on
SOCI library.

We can revert this change if we wish once SOCI updates to containerd v2.

Signed-off-by: David Son <davbson@amazon.com>