Changing bufio.Scanner to bufio.Reader to support large message
From the Issue #3343

As stated in the bufio Documentation, when using the bufio.Scanner, if the buffer exceed the max size of the token, the Scanner fail silently with an I/O error.

In the case of a container started with nerdctl, the main process using stdout/stderr hang forever when sending text bigger than 64kb (bufio.MaxScanTokenSize) without any newline character. (Discovered from a JSON string dumped in the logs). bufio documentation state to switch to a Reader instead when you do not have control over the size of the message.

Small test that is failing with the current master but working with this fix.

@apostasie recommend to check the other place in the code for the bufio.Scanner usage but I am not really a GO expert

Will trim any \n if it is not EOF else, will send the data no matter and exit in case of error

Signed-off-by: Bastien Roy-Mazoyer <bastien9222@hotmail.com>