Add support for explicitly setting --insecure-registry=false
As described in https://github.com/containerd/nerdctl/issues/3046 it is not currently possible to communicate
with localhost using TLS, as insecure-registry is forcefully set for localhost-ish addresses.
This PR does introduce support to differenciate between the default value of --insecure-registry when it is NOT
specified (which is false), and the fact that the user has explicitly passed the flag as `--insecure-registry=false`.

Being able to differenciate these two cases will enable us to treat differently:
`docker login localhost` (usual behavior, insecure set to true)
and
`docker login --insecure-regitry=false localhost` (new behavior, NOT setting insecure to true)

Implementation will follow in a subsequent PR that will rewrite login.go.

Signed-off-by: apostasie <spam_blackhole@farcloser.world>