Fix test to workaround docker credstore symlink fault
As described in https://github.com/containerd/nerdctl/issues/3413 there seems to be a fault
in docker credentials store that makes relative symlinks resolve to pwd instead of where they are.
This in turn will break our test if the current working directory is read-only (typical with Lima).
This changeset does Chdir to the parent temp directory so we can workaround that problem.

Signed-off-by: apostasie <spam_blackhole@farcloser.world>