Fix panic 'provided file is not a console'
As outlined in https://github.com/containerd/nerdctl/issues/3433, containerd/console will panic on a call to console.Current().
This patch provides a simple consoleutil wrapper that will return an error instead.

Note that part of https://github.com/containerd/nerdctl/pull/3300 is being reverted, as no longer necessary.

This patch does not try to be "smart" and does not check the status of stdin/out/err otherwise or if it is consistent with user provided flags,
but merely ensures we do not crash.

Signed-off-by: apostasie <spam_blackhole@farcloser.world>