Prevent concurrency on commit and remove
Some container operations require exclusive access to the container, and/or have no elegant way to deal with
concurrent container removal.
See for example: https://github.com/containerd/nerdctl/issues/1391

This PR adds a locking method that containers can use, backed by a store instance inside the container statedir,
and enforces exclusive locking for `Commit` and `Remove`.
There may be other container operations where we would like to apply the lock as well (to be evaluated).
Finally note that this locking mechanism is to prevent another instance of nerdctl from executing concurrently,
and does not offer guarantees that the container will not be manipulated by a different cli or otherwise change
state on its own.

Signed-off-by: apostasie <spam_blackhole@farcloser.world>