fix: Cleaning up orphaned directories and files when containers creation fails
When trying to create new containers, the following directories are
created based on the new container ID:

  - `/var/lib/nerdctl/1935db59/containers/<namespace>/<container id>`
  - `/var/lib/nerdctl/1935db59/etchosts/<namespace>/<container id>`

When containers with existing names are attempted to be created, the
processes fail.
However, in the events of failures, these directories are not cleaned up.

This issue is reported in the following:

  - https://github.com/containerd/nerdctl/issues/2993

This commit resolves the issue by cleaning up the mentioned directories
when containers creation fails.

Also, It has also been modified so that the following directory is also
cleaned up when the container is removed.

  - `/var/lib/nerdctl/1935db59/etchosts/<namespace>/<container id>`

Note that tests of logic to fix bug in Issue #2993 are added based on the
following testing principles.

  - https://github.com/containerd/nerdctl/blob/main/docs/testing/tools.md

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>