Do not print an error message on non-0 exec exit code
This was added with an earlier exec, and it is very confusing. the error had nothing to do with Containerd; it was the executable we ran inside the container that errored, and per task run convention we should set the Containerd exit code to the process's exit code and print no error.

Signed-off-by: Zou Nengren <zouyee1989@gmail.com>