fix: save multiple images for the same image id to a tar archive
Suppose we try to save multiple container images with the same image ID
but different image names into a tar archive using the nerdctl save
command.

When we then try to load container images from this tar archive using the
nerdctl load command, not all container images will be loaded.

This behavior is reported and the details are described in the following:

- https://github.com/containerd/nerdctl/issues/3806

Therefore, this PR resolves this issue.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>