test: Add a test for container access with 127.0.0.2 specified in -p in rootless mode
When running a rootless container, specifying an address such as 127.0.0.2
for the -p option, we will not be able to access the published container
through that address.

This behavior is reported in the following issue:

- https://github.com/containerd/nerdctl/issues/3539

This behavior is caused by the behavior in rootlesskit, and the following
pull request has been made to improve the behavior.

- https://github.com/rootless-containers/rootlesskit/pull/477

Therefore, this commit adds a test to ensure that the behavior of the
issue has been improved.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>