Remove golanci-lint action
Depending on the action means depending on a precompiled version of it that is incompatible with updated golang versions.
Alternatively, the action offers to build it from source, but unfortunately, the specific GOOS has to apply for both
the compilation and verification stages.

The action also performs additional steps (--verify) that are unnecessary and have been a source of issues.

Finally, golangci-lint version has to be maintained in two distinct places:
- local developer environment
- github CI side

Removing the action in favor of calling `make install-dev-tools` will fix these issues.

Signed-off-by: apostasie <spam_blackhole@farcloser.world>