test: update test logic in TestTestParseFlagPWithPlatformSpec
This test does not run in the rootfull Linux environment.
Also, there are some cases where the test does not run.
Therefore, this commit modifies all tests to perform the evaluation.

Note that when the host ports are not specified in the -p option of
`nerdctl run`, ports are randomly assigned, but it is not known what port
number will be assigned in each test.
Therefore, this commit modifies not to compare host ports.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>