fix: allow storing additional network info in network-config.json
When we specify the `-p` option with `nerdctl run`, etc., the port mapping
information is stored in the following `network-config.json`.

```
<DATAROOT>/<ADDRHASH>/containers/<NAMESPACE>/<CID>/network-config.json
```

However, the current implementation can only store port mapping
information and does not have the extensibility to store other
network-related information.

This point was feedback in the following PR.

- https://github.com/containerd/nerdctl/pull/4376

Therefore, this commit updates the `network-config.json` to allow for the
expansion of the information that can be stored.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>