rootful: reserve the ports on the host
When running in rootful mode, reserve the ports on the host
so that the ports appears on /proc/net/tcp.

This also prevents other processes from binding to the same ports.

Note that in rootless mode this is not necessary because
RootlessKit's port driver already reserves the ports.

See lima-vm/lima issue 4085

Similar patterns are used in Docker and Podman.
- moby/moby PR 48132
- containers/podman PR 23446

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>