fix: use private namespace for image build in private namespace
In the current implementation, tests that require nerdtest.Build do not
run against the specified namespace when a private namespace is
configured.

For example, with the following requirements, the test runs in a private
namespace, but the internal logic of nerdtest.Build does not take the
private namespace into account, causing buildctl or buildkit to be
unavailable:

```
testCase.Require = require.All(
    nerdtest.Private,
    nerdtest.Build,
)
```

This commit fixes the behavior so that when a private namespace is
specified, tests are executed considering that namespace instead of the
default namespace.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>