chore: improve the visibility of orphaned containers in logs
When executing the command on a compose file like the one below, orphaned
containers are displayed, but the current implementation does not show
the IDs of these orphaned containers, making them difficult to identify.

```
$ cat compose-full.yaml
services:
  test:
    image: docker.io/library/busybox:latest
    command: ["sleep", "infinity"]
  orphan:
    image: docker.io/library/busybox:latest
    command: ["sleep", "infinity"]

$ cat compose-orphan.yaml
services:
  test:
    image: docker.io/library/busybox:latest
    command: ["sleep", "infinity"]

$ sudo nerdctl compose -f compose-full.yaml up -d
...

$ sudo nerdctl compose -f compose-orphan.yaml down -v
...
WARN[0010] found 1 orphaned containers: [0x4000340000], you can run this command with the --remove-orphans flag to clean it up
...
```

Therefore, this commit modifies to display the IDs of orphaned containers.

Additionally, since there was other logic that performed similar displays,
this commit also modifies it in the same manner.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>